requirements.txt
fastapi
uvicorn
pydantic
mcp
fastmcp
typing-extensions
google-generativeai
python-dotenv
